Adds database configuration
