+++ server/.env.example
@@ -1,5 +1,12 @@
 # Define which port to use here
 PORT=
 
-#Define the database Uri here
-DATABASE_URI=
+# Define the database credentials here
+DB_NAME=
+DB_USER=
+DB_PASS=
+DB_HOST=
+DB_PORT=
+
+# Define the node environment to run in (default: "development")
+NODE_ENV=
